move utils into folder
